Reorganize CLI Menus into separate classes.
