+++ cli-ui/cli-menu-factory.js
@@ -1,14 +1,19 @@
-const {ActiveFastMenu, InactiveFastMenu, CurrentFastMenu, NoCurrentFastMenu, FastTypeOptionsMenu, AllPreviousFastsMenu} = require('./cli-menus');
+const ActiveFastMainMenu = require('./cli-menus/active-fast-menu');
+const InactiveFastMainMenu = require('./cli-menus/inactive-fast-menu');
+const CurrentFastInfoMenu = require('./cli-menus/current-fast-info-menu');
+const NoCurrentActiveFastMenu = require('./cli-menus/no-current-active-fast-menu');
+const FastTypeOptionsMenu = require('./cli-menus/fast-types-options-menu');
+const AllPreviousFastsInfoMenu = require('./cli-menus/show-all-previous-active-fasts');
 
 module.exports = CliMenuFactory = (type) => {
     switch (type) {
         case 'NO_ACTIVE_FAST' :
         {
-            return (new InactiveFastMenu()).menu;
+            return (new InactiveFastMainMenu()).menu;
         }
         case 'ACTIVE_FAST' :
         {
-            return (new ActiveFastMenu()).menu;
+            return (new ActiveFastMainMenu()).menu;
         }
         case 'FAST_TYPE_OPTIONS' :
         {
@@ -16,14 +21,14 @@ module.exports = CliMenuFactory = (type) => {
         }
         case 'CURRENT_FAST' :
         {
-            return (new CurrentFastMenu()).menu;
+            return (new CurrentFastInfoMenu()).menu;
         }
         case 'NO_CURRENT_FAST':
         {
-            return (new NoCurrentFastMenu()).menu;
+            return (new NoCurrentActiveFastMenu()).menu;
         }
         case 'ALL_PREVIOUS_FASTS': {
-            return (new AllPreviousFastsMenu()).menu;
+            return (new AllPreviousFastsInfoMenu()).menu;
         }
 
         default :
